+++ apps/web/core/providers.tsx
@@ -1,20 +1,18 @@
 'use client';
 
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { Provider } from 'jotai';
 
 import * as React from 'react';
 
+import { ReactQueryProvider } from './query-client';
 import { Services } from './services';
 import { ActionsStoreProvider } from './state/actions-store/actions-store-provider';
 import { ActiveProposalProvider } from './state/active-proposal-store';
 import { DiffProvider } from './state/diff-store';
+import { JotaiProvider } from './state/jotai-provider';
 import { LocalStoreProvider } from './state/local-store';
 import { StatusBarContextProvider } from './state/status-bar-store';
 import { WalletProvider } from './wallet';
-
-const queryClient = new QueryClient();
 
 interface Props {
   onConnectionChange: (type: 'connect' | 'disconnect', address: string) => Promise<void>;
@@ -23,8 +21,8 @@ interface Props {
 
 export function Providers({ children, onConnectionChange }: Props) {
   return (
-    <QueryClientProvider client={queryClient}>
-      <Provider>
+    <ReactQueryProvider>
+      <JotaiProvider>
         <WalletProvider onConnectionChange={onConnectionChange}>
           <Services.Provider>
             <ActionsStoreProvider>
@@ -38,9 +36,9 @@ export function Providers({ children, onConnectionChange }: Props) {
             </ActionsStoreProvider>
           </Services.Provider>
         </WalletProvider>
-      </Provider>
+      </JotaiProvider>
 
       <ReactQueryDevtools initialIsOpen={false} />
-    </QueryClientProvider>
+    </ReactQueryProvider>
   );
 }
