+++ apps/web/core/state/jotai-provider.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { Provider } from 'jotai';
+
+export function JotaiProvider({ children }: { children: React.ReactNode }) {
+  return <Provider>{children}</Provider>;
+}
