+++ apps/web/package.json
@@ -59,6 +59,7 @@
     "connectkit": "1.5.1",
     "csv": "^6.3.3",
     "dayjs": "^1.11.9",
+    "dexie": "^3.2.4",
     "diff": "^5.1.0",
     "effect": "2.0.0-next.48",
     "framer-motion": "^10.15.2",
